pyproject.toml
[project]
name = "shortfx"
version = "0.3.1"
description = "3000+ deterministic Python functions for AI agents — MCP server, JSON Schema, zero deps"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Javier Rodríguez", email = "jrodriguezga@outlook.com" },
]
keywords = [
    "ai-agent",
    "mcp",
    "ai-tools",
    "function-calling",
    "deterministic",
    "formulas",
    "excel",
    "llm",
    "openai",
    "langchain",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
    "Typing :: Typed",
]
dependencies = []

[project.urls]
Homepage = "https://github.com/jrodriguezgar/shortfx"
Repository = "https://github.com/jrodriguezgar/shortfx"
Issues = "https://github.com/jrodriguezgar/shortfx/issues"
Changelog = "https://github.com/jrodriguezgar/shortfx/blob/main/CHANGELOG.md"

[project.optional-dependencies]
mcp = ["mcp[cli]>=1.0.0"]
semantic = ["fastembed>=0.4.0", "truststore>=0.9.0"]
mcp-semantic = ["mcp[cli]>=1.0.0", "fastembed>=0.4.0", "truststore>=0.9.0"]
scipy = ["scipy>=1.10.0"]
dateutil = ["python-dateutil>=2.8.0"]
spellcheck = ["rapidfuzz>=3.0.0", "pyspellchecker>=0.7.0", "symspellpy>=6.7.0"]
similarity = ["textdistance>=4.5.0", "metaphone>=0.6", "python-Levenshtein>=0.21.0", "jellyfish>=1.0.0"]
all = [
    "mcp[cli]>=1.0.0",
    "fastembed>=0.4.0",
    "truststore>=0.9.0",
    "scipy>=1.10.0",
    "python-dateutil>=2.8.0",
    "rapidfuzz>=3.0.0",
    "pyspellchecker>=0.7.0",
    "symspellpy>=6.7.0",
    "textdistance>=4.5.0",
    "metaphone>=0.6",
    "python-Levenshtein>=0.21.0",
    "jellyfish>=1.0.0",
]

[project.scripts]
shortfx-mcp = "shortfx.mcp.server:main"

[dependency-groups]
dev = [
    "build>=1.2.0",
    "hatchling>=1.25.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-timeout>=2.4.0",
    "twine>=6.0.0",
]
docs = [
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shortfx"]

[tool.coverage.run]
source = ["shortfx"]
omit = [
    "shortfx/mcp/*",
    "shortfx/mcp_server.py",
    "shortfx/__main__.py",
    "shortfx/semantic_search.py",
    "shortfx/fxExcel/statistic_formulas.py",
    "shortfx/fxExcel/engineering_formulas.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]
